backend-django/requirements.txt
django==3.2
Pillow==8.2.0
psycopg2==2.8.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.7.0
channels-redis==2.4.2
channels==3.0.4